Revert "WIP: refactoring to v2"
This reverts commit 2a2cf67c035c640ad4173e943346db3ed3868346.
