+++ src/controllers/EmailVerification.js
@@ -10,12 +10,14 @@ const response = require("../utils/response");
 
 class EmailVerification {
   async verifyEmail(req, res) {
-    await EmailVerService.verifyEmail(req);
+    const data = await EmailVerService.verifyEmail(req);
 
     // return res.status(200).send(response("Email Address Verified", data));
     //hardcoding call back url to redirect back to microapi site
 
-    res.redirect("https://microapi.dev/accounts/signin");
+    if (data.success) {
+      res.redirect("https://microapi.dev/accounts/signin");
+    }
   }
 
   async resendVerification(req, res) {
